+++ rtos/pmsis/pmsis_bsp/camera/ov7670/ov7670.c
@@ -189,7 +189,7 @@ static inline int is_i2c_active()
 
 
 
-static void __ov7670_reg_write(ov7670_t *ov7670, unsigned int addr, uint8_t value)
+static void __ov7670_reg_write(ov7670_t *ov7670, uint8_t addr, uint8_t value)
 {
   if (is_i2c_active())
   {
@@ -201,7 +201,7 @@ static void __ov7670_reg_write(ov7670_t *ov7670, unsigned int addr, uint8_t valu
 
 
 
-static uint8_t __ov7670_reg_read(ov7670_t *ov7670, unsigned int addr)
+static uint8_t __ov7670_reg_read(ov7670_t *ov7670, uint8_t addr)
 {
   if (is_i2c_active())
   {
@@ -220,8 +220,8 @@ static uint8_t __ov7670_reg_read(ov7670_t *ov7670, unsigned int addr)
 
 static void __ov7670_init_regs(ov7670_t *ov7670)
 {
-    unsigned int i;
-    for(i=0; i<(sizeof(__ov7670_reg_init)/sizeof(ov7670_reg_init_t)); i++)
+    int32_t i;
+    for(i=0; i<(int32_t)(sizeof(__ov7670_reg_init)/sizeof(ov7670_reg_init_t)); i++)
     {
         __ov7670_reg_write(ov7670, __ov7670_reg_init[i].addr, __ov7670_reg_init[i].value);
     }
